Created a controllers folder and moved the logic inside the controller file
