+++ routes/users.js
@@ -1,57 +1,17 @@
 import express from 'express'
-import { v4 as uuidv4 } from 'uuid'
+import { createUser, getUsers, getUser, deleteUser, updateUser } from '../controllers/users.js'
 
 const router = express.Router()
 
-let users = []
+router.get('/', getUsers)
 
+router.post('/', createUser)
 
-// all routes here start with /users
-router.get('/', (req, res) => {
-  res.send(users)
-})
+router.get('/:id', getUser)
 
+router.delete('/:id', deleteUser)
 
-router.post('/', (req, res) => {
-  const user = req.body
-
-  users.push({ ...user, id: uuidv4() })
-
-  res.send(`User with the name ${user.firstName} added to the database`)
-})
-
-
-router.get('/:id', (req, res) => {
-  const { id } = req.params
-
-  const foundUser = users.find((user) => user.id === id)
-
-  res.send(foundUser)
-})
-
-
-router.delete('/:id', (req, res) => {
-  const { id } = req.params
-
-  users = users.filter((user) => user.id !== id)
-
-  res.send(`User with the id ${id} deleted from the database`)
-})
-
-
-router.patch('/:id', (req, res) => {
-  const { id } = req.params
-  const { firstName, lastName, age } = req.body;
-
-
-  const user = users.find((user) => user.id === id)
-
-  if (firstName) user.firstName = firstName
-  if (lastName) user.lastName = lastName
-  if (age) user.age = age
-
-  res.send(`User with the id ${id} has been updated`)
-})
+router.patch('/:id', updateUser)
 
 
 export default router
